# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal de compras - Farmácias Associadas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/programs": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "example": "teste"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "example": "123"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "example": "active"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProgramListResource"
                                            }
                                        },
                                        "links": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginationLinks"
                                                }
                                            ]
                                        },
                                        "meta": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Programs"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProgramCreatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Programa criado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProgramListResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProgramUpdatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Programa atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programs"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Programa removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProgramCreatorRequest": {
                "title": "Program form request",
                "required": [
                    "code",
                    "name",
                    "status"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                },
                "type": "object"
            },
            "ProgramUpdatorRequest": {
                "title": "Program update form request",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                },
                "type": "object"
            },
...
            "ProgramListResource": {
                "title": "Program Response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "code": {
                        "type": "string",
                        "example": "01"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                },
                "type": "object"
            },
...
            "UserStatus": {
                "type": "string",
                "enum": [
                    "ACTIVE",
                    "INACTIVE"
                ]
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=10"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=3"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "2"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "10"
                    },
                    "path": {
                        "type": "string",
                        "example": "http://api.test/api/users"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "20"
                    },
                    "to": {
                        "type": "integer",
                        "example": "3"
                    },
                    "total": {
                        "type": "integer",
                        "example": "10"
                    }
                },
                "type": "object"
            },
            "ValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": "Campo teste é obrigatório"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}